Micro-Manager/pyproject.toml
[project]
name = "micro-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.4",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "psycopg2-binary>=2.9.10",
]
